requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
cryptography==36.0.1
defusedxml==0.7.1
dj-rest-auth==2.2.2
Django==4.0.2
django-allauth==0.47.0
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-crontab==0.7.1
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
idna==3.3
namegenerator==1.0.6
oauthlib==3.2.0
pycparser==2.21
PyJWT==2.3.0
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
shortuuid==1.0.8
sqlparse==0.4.2
twilio==7.8.2
urllib3==1.26.8
